feat: :sparkles: Add reservation management to Battery class
The Battery class in the Rise.Domain.Boats namespace has been updated to include reservation management functionality. This includes the addition of a private list of reservations, as well as methods to add and remove reservations. The reservations are now exposed as a read-only collection property. This change improves the organization and encapsulation of reservation-related logic within the Battery class.